+++ src/main/java/ricedotwho/mf/Ticker.java
@@ -6,6 +6,8 @@ import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 import ricedotwho.mf.events.OnTimeEvent;
 import ricedotwho.mf.events.PacketEvent;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
@@ -13,7 +15,9 @@ import java.util.concurrent.TimeUnit;
 public class Ticker {
     static ScheduledExecutorService scheduler = Executors.newScheduledThreadPool(1);
     private static double ticksThisSecond = 0;
-    public static double tps = 20;
+    public static double instantTps = 20;
+    public static double tps = 20D;
+    private static List<Integer> pastTps = new ArrayList<>();
     public static boolean isServerTicking = false;
     @SubscribeEvent
     public void onServerTick(PacketEvent.ReceiveEvent event) {
@@ -26,11 +30,25 @@ public class Ticker {
     }
     @SubscribeEvent
     public void onUnload(WorldEvent.Unload event) {isServerTicking = false;}
+    public static void getTps() {
+        double total = 0D;
+        for (Integer secondTps : pastTps) {
+            total += secondTps;
+        }
+        tps = total / pastTps.size();
+    }
     static Runnable tpsCheck = new Runnable() {
         @Override
         public void run() {
             OnTimeEvent.Second.postAndCatch();
             tps = ticksThisSecond > 20 ? 20 : ticksThisSecond;
+            if(pastTps.size() < 5) {
+                pastTps.add((int) tps);
+            } else {
+                pastTps.remove(0);
+                pastTps.add((int) tps);
+            }
+            getTps();
             ticksThisSecond = 0;
         }
     };