+++ src/main/java/ricedotwho/mf/mining/MiningData.java
@@ -6,8 +6,6 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-
-//todo: make this cleaner. rn its 92.5% chatgpt
 public class MiningData {
     public static final Map<String, Integer> MINING_TALISMAN = new HashMap<>();
     public static final Map<String, Integer> PET_MS = new HashMap<>();
@@ -64,151 +62,6 @@ public class MiningData {
         MINING_HARDNESS.put("82:0", 5_600); // Clay
         MINING_HARDNESS.put("4:0", 5_600); // Cobblestone
     }
-
-    // GEMSTONE_STATS
-    public static final Map<String, Map<String, Map<String, Integer>>> GEMSTONE_STATS = new HashMap<>();
-    static {
-        Map<String, Map<String, Integer>> amberMap = new HashMap<>();
-
-        amberMap.put("COMMON", createGemstoneStats(4, 6, 10, 14, 20));
-        amberMap.put("UNCOMMON", createGemstoneStats(8, 10, 14, 20, 28));
-        amberMap.put("RARE", createGemstoneStats(12, 14, 20, 30, 40));
-        amberMap.put("EPIC", createGemstoneStats(16, 18, 28, 44, 60));
-        amberMap.put("LEGENDARY", createGemstoneStats(20, 24, 36, 58, 80));
-        amberMap.put("MYTHIC", createGemstoneStats(24, 30, 45, 75, 100));
-        amberMap.put("DIVINE", createGemstoneStats(28, 36, 54, 92, 120));
-
-        GEMSTONE_STATS.put("AMBER", amberMap);
-    }
-
-    private static Map<String, Integer> createGemstoneStats(int rough, int flawed, int fine, int flawless, int perfect) {
-        Map<String, Integer> map = new HashMap<>();
-        map.put("ROUGH", rough);
-        map.put("FLAWED", flawed);
-        map.put("FINE", fine);
-        map.put("FLAWLESS", flawless);
-        map.put("PERFECT", perfect);
-        return map;
-    }
-
-    // ARMOUR
-    public static final Map<String, Map<String, Integer>> ARMOUR = new HashMap<>();
-    static {
-        ARMOUR.put("DIVAN", createStatMap(80, 30));
-        ARMOUR.put("SORROW", createStatMap(50, 20));
-        ARMOUR.put("FLAME", createStatMap(25, 10));
-        ARMOUR.put("HEAT", createStatMap(10, 5));
-        ARMOUR.put("GOBLIN", createStatMap(10, 0));
-        ARMOUR.put("GLACITE", createStatMap(10, 0));
-    }
-
-    private static Map<String, Integer> createStatMap(int ms, int mf) {
-        Map<String, Integer> map = new HashMap<>();
-        map.put("ms", ms);
-        map.put("mf", mf);
-        return map;
-    }
-
-    // MODIFIERS
-    public static final Map<String, Map<String, Map<String, Integer>>> MODIFIERS = new HashMap<>();
-    static {
-        MODIFIERS.put("jaded", createModifierMap(
-                createStatMap(5, 5), createStatMap(12, 10), createStatMap(20, 15),
-                createStatMap(30, 20), createStatMap(45, 25), createStatMap(60, 30),
-                createStatMap(60, 30), createStatMap(60, 30), createStatMap(60, 30)
-        ));
-        MODIFIERS.put("auspicious", createModifierMap(
-                createStatMap(7, 0), createStatMap(14, 0), createStatMap(23, 0),
-                createStatMap(34, 0), createStatMap(45, 0), createStatMap(60, 0),
-                createStatMap(75, 0), createStatMap(75, 0), createStatMap(75, 0)
-        ));
-        MODIFIERS.put("lustrous", createModifierMap(
-                createStatMap(5, 5), createStatMap(10, 7), createStatMap(15, 9),
-                createStatMap(20, 12), createStatMap(25, 16), createStatMap(30, 20),
-                createStatMap(35, 25), createStatMap(35, 25), createStatMap(35, 25)
-        ));
-        MODIFIERS.put("ambered", createModifierMap(
-                createStatMap(25, 0), createStatMap(31, 0), createStatMap(38, 0),
-                createStatMap(46, 0), createStatMap(55, 0), createStatMap(65, 0),
-                createStatMap(75, 0), createStatMap(75, 0), createStatMap(75, 0)
-        ));
-        MODIFIERS.put("fleet", createModifierMap(
-                createStatMap(25, 0), createStatMap(31, 0), createStatMap(38, 0),
-                createStatMap(46, 0), createStatMap(55, 0), createStatMap(65, 0),
-                createStatMap(75, 0), createStatMap(75, 0), createStatMap(75, 0)
-        ));
-        MODIFIERS.put("mithraic", createModifierMap(
-                createStatMap(6, 0), createStatMap(12, 0), createStatMap(20, 0),
-                createStatMap(30, 0), createStatMap(40, 0), createStatMap(55, 0),
-                createStatMap(70, 0), createStatMap(70, 0), createStatMap(70, 0)
-        ));
-
-        MODIFIERS.put("stellar", createModifierMap(
-                createStatMap(3, 0), createStatMap(6, 0), createStatMap(9, 0),
-                createStatMap(12, 0), createStatMap(15, 0), createStatMap(20, 0),
-                createStatMap(25, 0), createStatMap(25, 0), createStatMap(25, 0)
-        ));
-        MODIFIERS.put("excellent", createModifierMap(
-                createStatMap(4, 0), createStatMap(8, 0), createStatMap(12, 0),
-                createStatMap(16, 0), createStatMap(20, 0), createStatMap(25, 0),
-                createStatMap(25, 0), createStatMap(25, 0), createStatMap(25, 0)
-        ));
-        MODIFIERS.put("fortunate", createModifierMap(
-                createStatMap(1, 1), createStatMap(2, 1), createStatMap(3, 1),
-                createStatMap(4, 2), createStatMap(6, 2), createStatMap(8, 3),
-                createStatMap(8, 3), createStatMap(8, 3), createStatMap(8, 3)
-        ));
-        MODIFIERS.put("sturdy", createModifierMap(
-                createStatMap(3, 0), createStatMap(6, 0), createStatMap(9, 0),
-                createStatMap(12, 0), createStatMap(15, 0), createStatMap(20, 0),
-                createStatMap(20, 0), createStatMap(20, 0), createStatMap(20, 0)
-        ));
-        MODIFIERS.put("glistening", createModifierMap(
-                createStatMap(0, 2), createStatMap(0, 3), createStatMap(0, 4),
-                createStatMap(0, 5), createStatMap(0, 6), createStatMap(0, 7),
-                createStatMap(0, 8), createStatMap(0, 8), createStatMap(0, 8)
-        ));
-    }
-
-    private static Map<String, Map<String, Integer>> createModifierMap(
-            Map<String, Integer> common, Map<String, Integer> uncommon, Map<String, Integer> rare,
-            Map<String, Integer> epic, Map<String, Integer> legendary, Map<String, Integer> mythic,
-            Map<String, Integer> divine, Map<String, Integer> special, Map<String, Integer> verySpecial) {
-        Map<String, Map<String, Integer>> map = new HashMap<>();
-        map.put("COMMON", common);
-        map.put("UNCOMMON", uncommon);
-        map.put("RARE", rare);
-        map.put("EPIC", epic);
-        map.put("LEGENDARY", legendary);
-        map.put("MYTHIC", mythic);
-        map.put("DIVINE", divine);
-        map.put("SPECIAL", special);
-        map.put("VERY", verySpecial);
-        return map;
-    }
-
-    // EQUIPMENT
-    public static final Map<String, Map<String, Integer>> EQUIPMENT = new HashMap<>();
-    static {
-        EQUIPMENT.put("MITHRIL_BELT", createStatMap(20, 5));
-        EQUIPMENT.put("MITHRIL_CLOAK", createStatMap(20, 5));
-        EQUIPMENT.put("MITHRIL_NECKLACE", createStatMap(20, 5));
-        EQUIPMENT.put("MITHRIL_GAUNTLET", createStatMap(20, 5));
-
-        EQUIPMENT.put("TITANIUM_BELT", createStatMap(30, 10));
-        EQUIPMENT.put("TITANIUM_CLOAK", createStatMap(30, 10));
-        EQUIPMENT.put("TITANIUM_NECKLACE", createStatMap(30, 10));
-        EQUIPMENT.put("TITANIUM_GAUNTLET", createStatMap(30, 10));
-
-        EQUIPMENT.put("JADE_BELT", createStatMap(30, 10));
-        EQUIPMENT.put("SAPPHIRE_CLOAK", createStatMap(30, 10));
-        EQUIPMENT.put("AMBER_NECKLACE", createStatMap(30, 10));
-        EQUIPMENT.put("AMETHYST_GAUNTLET", createStatMap(30, 10));
-
-        EQUIPMENT.put("DWARVEN_HANDWARMERS", createStatMap(45, 30));
-
-        EQUIPMENT.put("DIVAN_PENDANT", createStatMap(100, 25));
-    }
     public static final Map<String, Map<String, Integer>> ITEM_MS = new HashMap<>();
 
     static {
@@ -262,16 +115,6 @@ public class MiningData {
         stats.put("mf", mf);
         stats.put("bp", bp);
         ITEM_MS.put(itemName, stats);
-    }
-
-    public static final Map<String, Map<String, Integer>> DRILL_ENGINES = new HashMap<>();
-
-    static {
-        DRILL_ENGINES.put("MITHRIL_DRILL_ENGINE", createStatMap(50, 0));
-        DRILL_ENGINES.put("TITANIUM_DRILL_ENGINE", createStatMap(100, 0));
-        DRILL_ENGINES.put("RUBY_POLISHED_DRILL_ENGINE", createStatMap(150, 30));
-        DRILL_ENGINES.put("SAPPHIRE_POLISHED_DRILL_ENGINE", createStatMap(250, 50));
-        DRILL_ENGINES.put("AMBER_POLISHED_DRILL_ENGINE", createStatMap(400, 100));
     }
     static List<Item> MINING_ITEMS = new ArrayList<>();
     static {