+++ docs/CHANGELOG.md
@@ -14,7 +14,19 @@
  - Fixed a typo causing mixins to fail loading in v1.0.1
 
 (v1.1.1) — Bugfix (12/08/2024):
-- Hardstone mode now breaks hardstone in 2 ticks to avoid ghostblocks
-- Made sounds better
-- Added option to disable the exp gain sound
+ - Hardstone mode now breaks hardstone in 2 ticks to avoid ghostblocks
+ - Made sounds better
+ - Added option to disable the exp gain sound
+
+(v1.1.2) — Bugfix (14/08/2024):
+READ: THE MOD WILL NOT WORK UNTIL THE SERVER STARTS TICKING!!!
+This means the pingless mining WILL NOT WORK until the tps display says the tps is >0!
+Mining speed boost is known to get the wrong speed, will be fixed soon!
+- Attempting to fix block breaking inconsistency
+- Fixed mining speed boost
+- Fixed jumping and in water checks
+- Improved TPS display
+- Fixed Khazad-dûm not being detected as the Crystal Hollows due to weird formatting in scoreboard
+- Fixed the mod sometimes counting over 20 ticks per second, causing breaking to fail
+- Other misc renames